requirements.txt
streamlit==1.29.0
selenium==4.20.0
webdriver-manager==3.8.6
groq==1.5.0
python-dotenv==1.0.1
